+++ RELEASE_NOTES.md
@@ -1,10 +1,10 @@
 ## VPinFE Release Notes
 
 ### Summary
-This is the offical release of vpinfe v1.1.15.
+This is the offical release of vpinfe v1.1.16.
 
 ### What's New
-- New ARM release @evilwraith
+- Add global ini override to settings. aka the "-ini" option to launching
 
 ### Fixes
 None